alarms: convert log entry handling to use plugin extensions
Motivation:

The LogEntryHandler composes both logback appenders, like
the SMTP email appender, as well as an implementation of
a DAO store, in order to process alarms.

It would be preferable, however, to make this processing
extensible (for example, in order to call a RESTful service
to make a ticket entry in a ticket tracking system like
Service Now).

Modification:

Since we do not want to process raw events, but the converted
LogEntry object, implementing a logback appender for this
purpose is not optimal.  We have chosen instead to define
service provider interface.

The SPI exports the raw listener interfaces which other
modules may implement; but all such provider factories
must be defined and loaded inside the alarms Spring
application context.  The factories use the object
constructed by the service loader and do the
necessary configuration based on the beans available
in the application context.

Result:

Logging handler now also calls listeners which
represent specific implementations of the SPI.

Target: master
Acked-by: Tigran
Require-notes: yes
Require-book: yes