alarms: remove support for custom server-side regex alarm definitions
Motivation:

The alarms infrastructure provides for two types of alarms,
predefined (hardcoded), and custom.  The latter are generated
on the server-side from a set of regular expression rules.

The rationale for providing the second option was to give
administrators some flexibility in catching additional error
messages and promoting them to alarm status.  It was mostly
intended as a way of canvassing error output for eventual
inclusion as hardcoded/predefined alarms.

The usefulness of this feature seems to have run its course.
A poll among users elicited practically no response.  Because
supporting this infrastructure entails scalability complications
on the server side, we have decided the time has come to retire
it.

Modification:

This patch removes all areas of the code having to do with
the regex-type alarms definitions.  This includes command-
line and interpreter support and its related documentation.

A subsequent patch will complete this reduction by moving
the filtering entirely over to the client side.

Result:

A greatly reduced and simplified code-base for alarms.

Target: master
Require-notes: yes
Require-book: yes
Acked-by: Gerd