dcache-frontend: add 'pin' and 'unpin' activities to namespace resource POST method
Motivation:

After installation of 7.2 we have discovered a need for light weight fire and forget
pinning ability via REST API which is similar to "dccp -P -t -1" or pinning over NFS.

What has been noted is that QoS transition from tape -> disk+tape means changing file
state (NEARLINE -> ONLINE). Apparently in 5.2 it was still not the case. And this is indeed
a correct behavior.  Data handling group  at Fermilab is not ready for this (they are not
ready to manage space occupied by CUSTODUIAL/ONLINE replicas).
They prefer "fire and forget" method of "dccp -P -t -1" and polling
"are you there yet" or more recently subscribe to storage events.

Having looked at bulk service we concluded that it is not yet ready to mass scale production
operation. Although we believe that ultimately this is the proper substitute for SRM,
it is just not there  yet maturity wise.

Modification:

api/v1/namespace/path is extended by adding 'pin' and 'unpin'
activities to the POST method, utilizing a notify (fire-and-forget)
message to the PinManager.

Result:

Users have iterim functionality for staging as was available in
5.2 until the full-fledged Bulk service is ready for production.

We request a backport inasmuch as this represents, from the standpoint
of what was available in 5.2, a regression fix.

Target:  master
Request: 7.2
Request: 7.1
Request: 7.0
Request: 6.2
Requires-notes: yes
Requires-book: no
Patch: https://rb.dcache.org/r/13240/
Acked-by: Dmitry

Conflicts:
	modules/dcache-frontend/src/main/java/org/dcache/restful/resources/namespace/FileResources.java