pnfsmanager: make list scheduling behavior optional (selectable)
Motivation:

Commit d7653b9a01649ad30dd29541af467b95daefe6b8 has intrododuced
sequential listing and folding of list requests of the same directories.
This works very well in the environment where there are mutliple listing
requests to the same directories. But we discovered an edge case.
A site has a directory with 15M entries that is listed periodically.
Since it takes a lot of time to serve 15M listing, the list requests to
other directories that are dispatched to the same queue (due to modulo
of hashCode clash) are backing up even though the other threads are idle.

Modification:

Added dCache poroperty variable:

pnfsmanager.parallel-listing

that allows to restore previous behavior - listing requests are dispatched to
a single queue that is prrocessed by multiple parallel workers.

Result:

dCache admin can change behavior of listing scheduler.

Target: trunk
Request: 8.2, 9.2, 9.1, 9.0